chore(app): rename since well edits happen here
