+++ src/api/uic-inventory/src/components/Pages/index.jsx
@@ -1,6 +1,6 @@
 import Profile from './Profile';
 import ContactProgram from './ContactProgram';
-import CreateWell from './Wells/CreateWell';
+import CreateOrEditWell from './Wells/CreateOrEditWell';
 
 import CreateOrEditSite from './Sites/CreateOrEditSite';
 import AddSiteContacts from './Sites/AddSiteContacts';
@@ -10,4 +10,4 @@ export { SitesAndInventory, GenericLandingPage } from './Home';
 
 const Sites = { CreateOrEditSite, AddSiteContacts, AddSiteLocation };
 
-export { CreateWell, ContactProgram, Profile, Sites };
+export { CreateOrEditWell, ContactProgram, Profile, Sites };
